update: sperate jobprocessinglock to new file
