+++ crates/orchestrator/src/tests/config.rs
@@ -29,13 +29,12 @@ use crate::cli::queue::QueueValidatedArgs;
 use crate::cli::settlement::SettlementValidatedArgs;
 use crate::cli::snos::SNOSParams;
 use crate::cli::storage::StorageValidatedArgs;
-use crate::config::{
-    get_aws_config, Config, JobProcessingState, OrchestratorParams, ProcessingLocks, ProviderConfig, ServiceParams,
-};
+use crate::config::{get_aws_config, Config, OrchestratorParams, ProviderConfig, ServiceParams};
 use crate::data_storage::aws_s3::AWSS3ValidatedArgs;
 use crate::data_storage::{DataStorage, MockDataStorage};
 use crate::database::mongodb::MongoDBValidatedArgs;
 use crate::database::{Database, MockDatabase};
+use crate::helpers::{JobProcessingState, ProcessingLocks};
 use crate::queue::sqs::AWSSQSValidatedArgs;
 use crate::queue::{MockQueueProvider, QueueProvider};
 use crate::routes::{get_server_url, setup_server, ServerParams};
